refactor: browser communications manager
