+++ src/Packages/Passport/Runtime/Scripts/Browser.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0a4d272f8b8370d449eb9bd093a0f851
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
