+++ src/Packages/Passport/Runtime/Scripts/Core/BrowserCommunicationsManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 186c2fdd5529c7d4ead5b904dfca4cf6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
