+++ src/Packages/Passport/Runtime/Scripts/Passport.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
 using System.Threading.Tasks;
 using UnityEngine;
 using VoltstroStudios.UnityWebBrowser.Core;
@@ -9,6 +7,7 @@ using Newtonsoft.Json;
 using Immutable.Passport.Utility;
 using System.IO;
 using Immutable.Passport.Model;
+using Immutable.Passport.Core;
 
 namespace Immutable.Passport
 {
@@ -29,14 +28,13 @@ namespace Immutable.Passport
             private WebBrowserClient webBrowserClient = new();
         #endif
 
-        // Request ID to TaskCompletionSource
-        // Storing TaskCompletionSource as an object as C# doesn't support wildcards like TaskCompletionSource<Any>
-        // and using TaskCompletionSource<object> doesn't work. 
-        IDictionary<string, object> requestTaskMap = new Dictionary<string, object>();
-
         public event PassportReadyDelegate OnReady;
 
         private AuthManager auth = new();
+        private BrowserCommunicationsManager communicationsManager;
+
+        private Passport() {
+        }
 
         public static Passport Init() {
             if (Instance == null) {
@@ -53,8 +51,9 @@ namespace Immutable.Passport
         private void Start()
         {
             webBrowserClient.Init();
-            webBrowserClient.OnUnityPostMessage += OnUnityPostMessage;
             webBrowserClient.OnLoadFinish += OnLoadFinish;
+
+            communicationsManager = new BrowserCommunicationsManager(webBrowserClient);
         }
 
         private void Awake() {
@@ -87,8 +86,6 @@ namespace Immutable.Passport
                 webBrowserClient.OnLoadFinish -= OnLoadFinish;
             }
         }
-
-        #region Passport request
 
         /// <summary>
         /// Connects the user to Passport.
@@ -125,14 +122,14 @@ namespace Immutable.Passport
             GetImxProviderRequest request = new GetImxProviderRequest(u.idToken, u.accessToken, u.refreshToken, u.profile, u.etherKey);
             string data = JsonConvert.SerializeObject(request);
 
-            bool success = await createCallTask<bool>(PassportFunction.GET_IMX_PROVIDER, data);
+            bool success = await communicationsManager.Call<bool>(PassportFunction.GET_IMX_PROVIDER, data);
             if (!success) {
                 throw new PassportException("Failed to get IMX provider", PassportErrorType.WALLET_CONNECTION_ERROR);
             }
         }
 
         public Task<string?> GetAddress() {
-            return createCallTask<string?>(PassportFunction.GET_ADDRESS);
+            return communicationsManager.Call<string?>(PassportFunction.GET_ADDRESS);
         }
 
         public void Logout() {
@@ -158,107 +155,7 @@ namespace Immutable.Passport
         }
 
         public Task<string?> SignMessage(string message) {
-            return createCallTask<string?>(PassportFunction.SIGN_MESSAGE, message);
+            return communicationsManager.Call<string?>(PassportFunction.SIGN_MESSAGE, message);
         }
-
-        private Task<T> createCallTask<T>(string fxName, string? data = null) {
-            var t = new TaskCompletionSource<T>();
-            string requestId = call(fxName, data);
-
-            // Add task completion source to the map so we can return the reponse
-            requestTaskMap.Add(requestId, t);
-
-            return t.Task;
-        }
-
-        private string call(string fxName, string? data = null) {
-            string requestId = Guid.NewGuid().ToString();
-            Debug.Log($"{TAG} Call {fxName} (request ID: {requestId})");
-
-            Request request = new Request(fxName, requestId, data);
-            string requestJson = JsonConvert.SerializeObject(request).Replace("\\", "\\\\").Replace("\"", "\\\"");
-
-            // Call the function on the JS side
-            string js = @$"callFunction(""{requestJson}"")";
-            webBrowserClient.ExecuteJs(js);
-
-            return requestId;
-        }
-
-        #endregion
-
-        #region Passport response
-
-        private void OnUnityPostMessage(string message) {
-            Debug.Log($"[Unity] OnUnityPostMessage: {message}");
-            handleResponse(message);
-        }
-
-        private void handleResponse(string message) {
-            Response? response = JsonUtility.FromJson<Response>(message);
-
-            // Check if the reponse returned is valid and the task to return the reponse exists
-            if (response == null || response.responseFor == null || response.requestId == null)
-                return;
-
-            string requestId = response.requestId;
-            PassportException? exception = parseError(response);
-
-            if (requestTaskMap.ContainsKey(requestId)) {
-                switch (response.responseFor) {
-                    case PassportFunction.GET_ADDRESS:
-                        notifyRequestResult<string?>(requestId, JsonUtility.FromJson<AddressResponse>(message)?.address, exception);
-                        break;
-                    case PassportFunction.GET_IMX_PROVIDER:
-                        notifyRequestResult<bool>(requestId, response.success, exception);
-                        break;
-                    case PassportFunction.SIGN_MESSAGE:
-                        notifyRequestResult<string?>(requestId, JsonUtility.FromJson<SignMessageResponse>(message)?.result, exception);
-                        break;
-                    default:
-                        break;
-                }
-            } else {
-                throw new PassportException($"No TaskCompletionSource for request id {requestId} found.");
-            }
-        }
-
-        private PassportException? parseError(Response? response) {
-            if (response != null && (response.success == false || response.error != null)) {
-                // Failed or error occured
-                try {
-                    if (response.errorType != null) {
-                        PassportErrorType type = (PassportErrorType) System.Enum.Parse(typeof(PassportErrorType), response.errorType);
-                        return new PassportException(response.error, type);
-                    }
-                } catch (Exception ex) {
-                    Debug.Log($"{TAG} Parse passport type error: {ex.Message}");
-                }
-                return new PassportException(response.error);
-            } else {
-                // No error
-                return null;
-            }
-        }
-        
-        private void notifyRequestResult<T>(string requestId, T result, PassportException? e)
-        {
-            TaskCompletionSource<T?> completion = requestTaskMap[requestId] as TaskCompletionSource<T?>;
-            try {
-                if (e != null) {
-                    if (!completion.TrySetException(e))
-                        throw new PassportException($"Unable to set exception for for request id {requestId}. Task has already been completed.");
-                } else {
-                    if(!completion.TrySetResult(result))
-                        throw new PassportException($"Unable to set result for for request id {requestId}. Task has already been completed.");
-                }
-            } catch (ObjectDisposedException exception) {
-                throw new PassportException($"Task for request id {requestId} has already been disposed and can't be updated.");
-            }
-
-            requestTaskMap.Remove(requestId);
-        }
-
-        #endregion
     }
 }
